refactor: rename /pkg to /internal
