+++ internal/paths/paths_test.go
