+++ internal/teeth/rawmetadata.go
